pyproject.toml
[project]
name = "camera-calibration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.6",
    "opencv-python>=4.12.0.88",
    "pyyaml>=6.0.3",
]
